+++ write_json.cpp
@@ -247,7 +247,17 @@ struct lonlat {
 	}
 };
 
-void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y, bool comma, bool name, bool zoom, bool dropped, unsigned long long index, long long sequence, long long extent, bool complain, json_writer &state) {
+void write_coords(json_writer &state, lonlat const &ll, double scale) {
+	if (scale == 0) {
+		state.json_write_float(ll.lon);
+		state.json_write_float(ll.lat);
+	} else {
+		state.json_write_number(ll.x / scale);
+		state.json_write_number(ll.y / scale);
+	}
+}
+
+void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y, bool comma, bool name, bool zoom, bool dropped, unsigned long long index, long long sequence, long long extent, bool complain, json_writer &state, double scale) {
 	for (size_t f = 0; f < layer.features.size(); f++) {
 		mvt_feature const &feat = layer.features[f];
 
@@ -359,9 +369,9 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 			long long py = feat.geometry[g].y;
 
 			if (op == VT_MOVETO || op == VT_LINETO) {
-				long long scale = 1LL << (32 - z);
-				long long wx = scale * x + (scale / layer.extent) * px;
-				long long wy = scale * y + (scale / layer.extent) * py;
+				long long wscale = 1LL << (32 - z);
+				long long wx = wscale * x + (wscale / layer.extent) * px;
+				long long wy = wscale * y + (wscale / layer.extent) * py;
 
 				double lat, lon;
 				projection->unproject(wx, wy, 32, &lon, &lat);
@@ -380,8 +390,7 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 				state.json_write_string("coordinates");
 
 				state.json_write_array();
-				state.json_write_float(ops[0].lon);
-				state.json_write_float(ops[0].lat);
+				write_coords(state, ops[0], scale);
 				state.json_end_array();
 			} else {
 				state.json_write_string("type");
@@ -392,8 +401,7 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 
 				for (size_t i = 0; i < ops.size(); i++) {
 					state.json_write_array();
-					state.json_write_float(ops[i].lon);
-					state.json_write_float(ops[i].lat);
+					write_coords(state, ops[i], scale);
 					state.json_end_array();
 				}
 
@@ -416,8 +424,7 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 
 				for (size_t i = 0; i < ops.size(); i++) {
 					state.json_write_array();
-					state.json_write_float(ops[i].lon);
-					state.json_write_float(ops[i].lat);
+					write_coords(state, ops[i], scale);
 					state.json_end_array();
 				}
 
@@ -435,8 +442,7 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 					if (ops[i].op == VT_MOVETO) {
 						if (sstate == 0) {
 							state.json_write_array();
-							state.json_write_float(ops[i].lon);
-							state.json_write_float(ops[i].lat);
+							write_coords(state, ops[i], scale);
 							state.json_end_array();
 
 							sstate = 1;
@@ -445,16 +451,14 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 							state.json_write_array();
 
 							state.json_write_array();
-							state.json_write_float(ops[i].lon);
-							state.json_write_float(ops[i].lat);
+							write_coords(state, ops[i], scale);
 							state.json_end_array();
 
 							sstate = 1;
 						}
 					} else {
 						state.json_write_array();
-						state.json_write_float(ops[i].lon);
-						state.json_write_float(ops[i].lat);
+						write_coords(state, ops[i], scale);
 						state.json_end_array();
 					}
 				}
@@ -570,13 +574,11 @@ void layer_to_geojson(mvt_layer const &layer, unsigned z, unsigned x, unsigned y
 				for (size_t j = 0; j < rings[i].size(); j++) {
 					if (rings[i][j].op != VT_CLOSEPATH) {
 						state.json_write_array();
-						state.json_write_float(rings[i][j].lon);
-						state.json_write_float(rings[i][j].lat);
+						write_coords(state, rings[i][j], scale);
 						state.json_end_array();
 					} else {
 						state.json_write_array();
-						state.json_write_float(rings[i][0].lon);
-						state.json_write_float(rings[i][0].lat);
+						write_coords(state, rings[i][j], scale);
 						state.json_end_array();
 					}
 				}